server: Split up original server.py into different sections.
server.py originally had the web application end points as well as all
of the taste_preference helper functions. They have now been split up
which should additionally allow for the inclusion of tests and easier
module management.

created: cred.py
created: taste_pref.py
modified: server.py
