+++ cred.py
@@ -1,0 +1,10 @@
+######## Credential Checking ########
+# TODO(kbona@): Work out credential checking.
+# Input:
+#  - (string) user's ID
+# Output:
+#  - (string) error
+def checkID(id):
+    return ""
+
+
